feat: added requested changes
